V1.1
Updated the help to have copyright and version numbers. Swapped the
fields around in the Compute_Pitch_And_Heading and Adjust_For_Pitch to
be consistent with Compute_Heading and The Compute_Pitch.

Added Option for entering the degree offset for Adjust_Pitch, it was
hardwired to 90 degrees in V1.00